requirements.txt
cachetools == 5.2.0
python-dotenv == 0.21.0
requests == 2.28.0
pillow == 11.1.0